ref(metadata): apply changes required from code review
