+++ api_metadata_test.go
@@ -2,8 +2,10 @@ package blockfrost_test
 
 import (
 	"context"
+	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
+	"path/filepath"
 	"testing"
 	"time"
 
@@ -26,20 +28,12 @@ func TestResourceMetadataTxLabels(t *testing.T) {
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`[
-					{
-					"label": "1990",
-					"cip10": null,
-					"count": "1"
-					},
-					{
-					"label": "1967",
-					"cip10": "nut.link metadata oracles registry",
-					"count": "3"
-					}
-					]`,
-			))
+			fp := filepath.Join(testdata, "json", "metadata", "metadata.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
 		}),
 	)
 	defer s.Close()
@@ -67,11 +61,11 @@ func TestResourceMetadataTxContentInJSON(t *testing.T) {
 	inputLabel := "1990"
 	expectedElement := blockfrost.MetadataTxContentInJSON{
 		TxHash: "257d75c8ddb0434e9b63e29ebb6241add2b835a307aa33aedba2effe09ed4ec8",
-		JSONMetadata: blockfrost.JSONMetadata{
-			Adausd: []blockfrost.Adausd{
-				{
-					Value:  "0.10409800535729975",
-					Source: "ergoOracles",
+		JSONMetadata: map[string]interface{}{
+			"ADAUSD": []interface{}{
+				map[string]interface{}{
+					"value":  "0.10409800535729975",
+					"source": "ergoOracles",
 				},
 			},
 		},
@@ -79,32 +73,13 @@ func TestResourceMetadataTxContentInJSON(t *testing.T) {
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`[
-					{
-						"tx_hash": "257d75c8ddb0434e9b63e29ebb6241add2b835a307aa33aedba2effe09ed4ec8",
-						"json_metadata": {
-							"ADAUSD": [
-								{
-									"value": "0.10409800535729975",
-									"source": "ergoOracles"
-								}
-							]
-						}
-					},
-					{
-						"tx_hash": "e865f2cc01ca7381cf98dcdc4de07a5e8674b8ea16e6a18e3ed60c186fde2b9c",
-						"json_metadata": {
-							"ADAUSD": [
-								{
-									"value": "0.15409850555139935",
-									"source": "ergoOracles"
-								}
-							]
-						}
-					}
-				]`,
-			))
+			fp := filepath.Join(testdata, "json", "metadata", "metadata_in_json.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
+
 		}),
 	)
 	defer s.Close()
@@ -125,70 +100,6 @@ func TestResourceMetadataTxContentInJSON(t *testing.T) {
 
 }
 
-func TestResourceMetadataTxContentInJSONRaw(t *testing.T) {
-	t.Parallel()
-
-	inputLabel := "1990"
-	expectedElement := blockfrost.MetadataTxContentInJSONRaw{
-		TxHash: "257d75c8ddb0434e9b63e29ebb6241add2b835a307aa33aedba2effe09ed4ec8",
-		JSONMetadata: map[string]interface{}{
-			"ADAUSD": []interface{}{
-				map[string]interface{}{
-					"value":  "0.10409800535729975",
-					"source": "ergoOracles",
-				},
-			},
-		},
-	}
-	s := httptest.NewServer(
-		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
-			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`[
-					{
-						"tx_hash": "257d75c8ddb0434e9b63e29ebb6241add2b835a307aa33aedba2effe09ed4ec8",
-						"json_metadata": {
-							"ADAUSD": [
-								{
-									"value": "0.10409800535729975",
-									"source": "ergoOracles"
-								}
-							]
-						}
-					},
-					{
-						"tx_hash": "e865f2cc01ca7381cf98dcdc4de07a5e8674b8ea16e6a18e3ed60c186fde2b9c",
-						"json_metadata": {
-							"ADAUSD": [
-								{
-									"value": "0.15409850555139935",
-									"source": "ergoOracles"
-								}
-							]
-						}
-					}
-				]`,
-			))
-		}),
-	)
-	defer s.Close()
-	api, err := blockfrost.NewAPIClient(
-		blockfrost.APIClientOptions{Server: s.URL},
-	)
-	if err != nil {
-		t.Fatalf(err.Error())
-	}
-	q := blockfrost.APIPagingParams{
-		Count: 1,
-	}
-	accountArray, err := api.MetadataTxContentInJSONRaw(context.TODO(), inputLabel, q)
-	if err != nil {
-		t.Fatalf(err.Error())
-	}
-	assert.Equal(t, expectedElement, accountArray[0])
-
-}
-
 func TestResourceMetadataTxContentInCBOR(t *testing.T) {
 	t.Parallel()
 
@@ -200,22 +111,13 @@ func TestResourceMetadataTxContentInCBOR(t *testing.T) {
 	s := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 			time.Sleep(20 * time.Millisecond)
-			w.Write([]byte(
-				`[
-					{
-						"tx_hash": "257d75c8ddb0434e9b63e29ebb6241add2b835a307aa33aedba2effe09ed4ec8",
-						"cbor_metadata": null
-					},
-					{
-						"tx_hash": "e865f2cc01ca7381cf98dcdc4de07a5e8674b8ea16e6a18e3ed60c186fde2b9c",
-						"cbor_metadata": null
-					},
-					{
-						"tx_hash": "4237501da3cfdd53ade91e8911e764bd0699d88fd43b12f44a1f459b89bc91be",
-						"cbor_metadata": "\\xa100a16b436f6d62696e6174696f6e8601010101010c"
-					}
-				]`,
-			))
+			fp := filepath.Join(testdata, "json", "metadata", "metadata_in_cbor.json")
+			file, err := ioutil.ReadFile(fp)
+			if err != nil {
+				t.Fatalf("an error ocurred while trying to read json test file %s", fp)
+			}
+			w.Write(file)
+
 		}),
 	)
 	defer s.Close()
